Revert "Convert more drop-ins into early ones"
This reverts commit 6b8dde1d4f867aff713af6d6830510a84fad58d2.

The risk of conflict with the unit files shipped by systemd is very low. Or to
say this differently, users are unlikely to override some settings by rewriting
the whole unit file. And since early drop-ins are deprecated, let's not rely on
them for stuff that they should be not needed.