user@.service: Disable controllers by default
The switch to cgroup v2 means that controller delegation is permitted for
any user (not only uid=0) within systemd.
Performance penalties are still measurable (as of ~v6.4 kernel), so we
want to keep controllers still disabled.

Change the delegation from straight 'no' into an empty list of
controllers (relies on support of list implemented since v236).

Extend the drop-in to any user's user@.service, not only user@0.service.

The effect of the change on unified setups is disabling of controllers
also for non-privileged user instances and the effect is void on hybrid
setups (non-privileged services are not allowed delegation implicitly
because of cgroup v1).

A possible alternative would have been to DisableControllers= on
ancestral units of user@.services. But: a) it needs explicit list of all
of them, b) we only want to affect the default (without preventing lazy
enablement of controllers when user e.g. sets up a memory limit on
user@uid.service).

[mkoutny: fixes jsc#PED-2276]